Introduce outdated PHP notice, deactivation.
